requirements.txt
asgiref
beautifulsoup4
celery
emoji
numpy
Django
djangorestframework
django-allauth==0.54.0
django-auditlog
django-snowflake
django-celery-beat
django-snowflake
django-cors-headers
dj-rest-auth
django-softdelete
django-role-permissions
djangorestframework-simplejwt
google-api-python-client
google-cloud-dialogflow-cx
google-auth
google-auth-oauthlib
gunicorn
instagrapi
mailchimp-marketing
openpyxl
pandas
Pillow
pre-commit
PyJWT
pytz
psycopg2-binary
redis
requests
selenium
sentry-sdk[django]
scrapy
sqlparse
coverage
robotframework
webdriver-manager



